Add files via upload
stopped videostream from recording when not needed (used file check)

+++ videostream_loop.py
@@ -77,6 +77,9 @@ class FERServer(asynchat.async_chat):
     def run_fer_loop(self, frame_cap = None, width=112, height=112, cam_port=0):
         class_names = ['Neutral', 'Happy', 'Sad', 'Surprise', 'Fear', 'Disgust', 'Angry'] 
         saveImagesPath = "./dataFolder/"
+        # the stop file tells the server NOT to save images (if it exists)
+        stopSavingImagesPath = saveImagesPath + 'STOP.txt'
+        
         framePrefix = "frame_" # name of images saved until made into video
         
         print("Loading camera. This takes a few seconds..")
@@ -93,16 +96,17 @@ class FERServer(asynchat.async_chat):
                 
                 #save original webcam image
                 try:
-                    numSuf = 0
-                    stNumSuf = str(numSuf).zfill(4)
-                    framePrefix = str(time.time())+"_"
-                    #imageName = saveImagesPath+framePrefix+stNumSuf+'.png'
-                    imageName = saveImagesPath+framePrefix+'.png'
-                    while(os.path.isfile(imageName)): # in case of name clash
-                        numSuf+=1
-                        stNumSuf = '_'+str(numSuf).zfill(4)
-                        imageName = saveImagesPath+framePrefix+stNumSuf+'.png' 
-                    cv2.imwrite(imageName, frame)
+                    if not os.path.exists(stopSavingImagesPath):
+                        numSuf = 0
+                        stNumSuf = str(numSuf).zfill(4)
+                        framePrefix = str(time.time())+"_"
+                        #imageName = saveImagesPath+framePrefix+stNumSuf+'.png'
+                        imageName = saveImagesPath+framePrefix+'.png'
+                        while(os.path.isfile(imageName)): # in case of name clash
+                            numSuf+=1
+                            stNumSuf = '_'+str(numSuf).zfill(4)
+                            imageName = saveImagesPath+framePrefix+stNumSuf+'.png' 
+                        cv2.imwrite(imageName, frame)
                 except Exception as e:
                     print(e)
